Revert "Treat XPASS status as a test failure"
For its immediate goal, this facilitated identifying the XPASS
status of TestSubmodule.test_rename and verifying that the revised
xfail condition eliminated it.

In the test output, XPASS is written as a failure, with strict
xfail behavior noted. Contributors less familiar with marking tests
xfail may mistake this to mean some behavior of the code under test
was broken. So if strict_xfail is to be enabled, it might be best
to do that in a pull request devoted to it, and maybe even to add
to the docs, about how to recognize and handle newly xpassing
tests.

This reverts commit 0f8cd4ce4a7f942d793a8670257f8738ab18b757.