Put back WSL on Windows CI
So the Windows test jobs can run the three out of four hook tests
that are able to pass with WSL bash, and so that we are able to
observe the expected failure on the one that still fails.

This undoes the removal of the setup-wsl step in d5ed266, now that
the test suite's bash.exe status checking logic, used for xfail
marks (and also usable in debugging), is internationalized. The CI
step was omitted during some of this process, to make sure
WinBashStatus would still detect the specific situation on the
GitHub Actions runner that occurs when no WSL distro is available.

(This commit thus has much of the same relationship to d5ed266
as 2875ffa had to cabb572.)