Restore longer extra time for Windows, and add for macOS
This undoes the testing breakage of test_blocking_lock_file for
Windows, and gives macOS the same amount of extra time as Windows.

I had expected that macOS, even though it needs more time than
Ubuntu, might not need as much extra time as Windows. That turned
out not to be the case, in the limited testing done so far on CI.
Windows, while still slower than Ubuntu and still too slow to
reliably pass the test_blocking_lock_file test, was usually faster
than macOS and passed more often.

These relative timings may turn out not to be a trend and only to
apply to the current GHA runners. That's probably okay since the
adjustment for macOS wasn't present before and is being added to
allow newly introduced macOS CI test jobs to pass.

That is all in regard to the very specific issue of the extra time
required for the test_blocking_lock_file test after the lock, which
makes an assertion about that not taking too long. Regarding the
overall time of entire test jobs, macOS 13 seems to have usually
been a little faster than macOS 12, so it is retained. Unlike
"macos-latest", which currently is macOS 12, "macos-13" will never
refer to a later version of the operating system, so the version
given in the workflow should be revisited later, at or after the
time "macos-latest" becomes a synonym of "macos-13".