Compare types in test_orig_head with "is"
This performs an exact type comparison with "is" instead of "==".
Most (maybe all?) of the rest of the places where "==" was used
have been changed previously, but this on fell through the cracks.
Like the None object, type objects use reference-based equality
comparison, and the idiomatic ways to check against them are "is"
for exactly comparisons and isinstance/issubclass otherwise.