Avoid making it look like kill_process works on Windows
This changes the code in Git.execute's local kill_process function,
which it uses as the timed callback for kill_after_timeout, to
remove code that is unnecessary because kill_process doesn't
support Windows, and to avoid giving the false impression that its
code could be used unmodified on Windows without serious problems.

- Raise AssertionError explicitly if it is called on Windows. This
  is done with "raise" rather than "assert" so its behavior doesn't
  vary depending on "-O".

- Don't pass process creation flags, because they were 0 except on
  Windows.

- Don't fall back to SIGTERM if Python's signal module doesn't know
  about SIGKILL. This was specifically for Windows which has no
  SIGKILL.

See #1756 for discussion.