Don't install black on Cygwin
As of black 23.12.0, installing black builds a wheel on Cygwin,
downloading dependencies for the "d" extra (i.e., black[d]), which
provides dependencies for running the blackd daemon. This includes
dependencies, such as asyncio, whose compilation fails on CI as it
is currently set up, fails under some common Cygwin setups (since
it needs a C compiler), and even if successful would cause the
Cygwin test workflow to take significantly longer to run.