Replace xfail with gc.collect in TestSubmodule.test_rename
Like the xfail was, this is conditional, being done only in the
specific situation the PermissionError occurs. Besides that it does
not always run even on Windows (only in 3.12 and later), this
resembles various other conditional and non-conditional gc.collect
calls.

It had previously appeared to me that two calls to gc.collect were
required, but I am unable to reproduce that. It may have been
specific to how I was running it on my system at that time. The
need for only one call may have been brought about by changes to
the code in the mean time, but I have tested that only one call
appears required even without the changes in #1765.